requirements.txt
db-dtypes
functions-framework
google-cloud-bigquery
pandas